+++ SailMate-app/Backend/src/main/java/group12/Backend/entity/Ticket.java
@@ -1,3 +1,4 @@
+// Ticket.java
 package group12.Backend.entity;
 
 import java.time.LocalDateTime;
@@ -22,7 +23,7 @@ public class Ticket {
     private Integer passengerCount;
     
     @Column(name = "total_price", nullable = false)
-    private Integer totalPrice;
+    private Double totalPrice;   // Changed from Integer to Double
     
     @Column(name = "ticket_class", nullable = false)
     private String ticketClass;
@@ -57,7 +58,7 @@ public class Ticket {
     public Ticket() {
     }
     
-    public Ticket(String ticketID, Integer voyageId, Integer passengerCount, Integer totalPrice, 
+    public Ticket(String ticketID, Integer voyageId, Integer passengerCount, Double totalPrice, 
                  String ticketClass, String selectedSeats, String userId, String ticketData) {
         this.ticketID = ticketID;
         this.voyageId = voyageId;
@@ -102,11 +103,11 @@ public class Ticket {
         this.passengerCount = passengerCount;
     }
 
-    public Integer getTotalPrice() {
+    public Double getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(Integer totalPrice) {
+    public void setTotalPrice(Double totalPrice) {
         this.totalPrice = totalPrice;
     }
